# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk sistem kasir dan antrean dapur.\nKasir membuat pesanan, koki memprosesnya secara realtime.",
        "title": "OrderFlow API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/reports/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ringkasan penjualan untuk satu tanggal. Pesanan yang dibatalkan tetap dihitung sebagai aktivitas, tetapi tidak masuk perhitungan omzet. Hanya kasir yang boleh mengakses.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Laporan penjualan harian",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tanggal laporan, format YYYY-MM-DD. Kosong berarti hari ini",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Jumlah menu terlaris yang ditampilkan",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DailyReport"
                        }
                    },
                    "403": {
                        "description": "Bukan kasir",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Format tanggal salah atau tanggal di masa depan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.DailyReport": {
            "type": "object",
            "properties": {
                "cancelled_orders": {
                    "type": "integer"
                },
                "cash_revenue": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "gross_revenue": {
                    "type": "integer"
                },
                "net_revenue": {
                    "type": "integer"
                },
                "non_cash_revenue": {
                    "type": "integer"
                },
                "top_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TopItem"
                    }
                },
                "total_discount": {
                    "type": "integer"
                },
                "total_orders": {
                    "type": "integer"
                }
            }
        },
...
        "models.TopItem": {
            "type": "object",
            "properties": {
                "menu_item_id": {
                    "type": "integer"
                },
                "menu_name": {
                    "type": "string"
                },
                "quantity_sold": {
                    "type": "integer"
                },
                "revenue": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}